Refactor imports and restructure files in Iterable module: consolidate exports, update paths for consistency, and enhance organization
