+++ src/__tests__/Iterable.test.ts
@@ -17,7 +17,7 @@ import {
   IterableEventName,
   IterableLogLevel,
 } from '..';
-import { IterableLogger } from '../IterableLogger';
+import { IterableLogger } from '../core/classes/IterableLogger';
 
 describe('Iterable', () => {
   beforeEach(() => {
