+++ src/core/classes/IterableAction.ts
