+++ src/core/classes/IterableActionContext.ts
@@ -1,4 +1,4 @@
-import type { IterableActionSource } from './enums';
+import type { IterableActionSource } from 'core/enums';
 import { IterableAction } from './IterableAction';
 
 /**
