+++ src/core/classes/IterableAttributionInfo.ts
