+++ src/core/classes/IterableAuthResponse.ts
