+++ src/core/classes/IterableCommerceItem.ts
