+++ src/core/classes/IterableEdgeInsets.ts
