+++ src/core/classes/IterableLogger.ts
