+++ src/core/classes/IterableUtil.ts
