+++ src/core/enums/IterableAuthResponseResult.ts
