+++ src/core/enums/IterableDataRegion.ts
