+++ src/core/index.ts
@@ -1,3 +1,3 @@
 export * from './classes';
 export * from './enums';
-export * from './types';
+export * from './hooks';
