+++ src/inApp/classes/IterableInboxMetadata.ts
