+++ src/inApp/classes/index.ts
@@ -1,0 +1,5 @@
+export * from './IterableHtmlInAppContent';
+export * from './IterableInAppManager';
+export * from './IterableInAppMessage';
+export * from './IterableInAppTrigger';
+export * from './IterableInboxMetadata';
