+++ src/inApp/types/IterableInAppContent.ts
@@ -1,4 +1,4 @@
-import type { IterableInAppContentType } from '../enums';
+import type { IterableInAppContentType } from 'inApp/enums';
 
 // TODO: Add description
 export interface IterableInAppContent {
