+++ src/inbox/components/index.ts
@@ -1,3 +1,5 @@
 export * from './IterableInbox';
 export * from './IterableInboxEmptyState';
 export * from './IterableInboxMessageCell';
+export * from './IterableInboxMessageDisplay';
+export * from './IterableInboxMessageList';
