+++ src/inbox/index.ts
@@ -1,2 +1,3 @@
+export * from './classes';
 export * from './components';
 export * from './types';
