+++ tsconfig.json
@@ -3,6 +3,8 @@
     "rootDir": ".",
     "paths": {
       "@iterable/react-native-sdk": ["./src/index"],
+      "core/*": ["./src/core/*"],
+      "core": ["./src/core/index"],
       "inbox/*": ["./src/inbox/*"],
       "inbox": ["./src/inbox/index"],
       "inApp/*": ["./src/inApp/*"],
